requirements.txt
fastapi
uvicorn
paypal-checkout-serversdk
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
python-multipart
